requirements.txt

fastapi
uvicorn
jinja2
reportlab
google-cloud-vision
google-cloud-storage
python-multipart
